+++ common/src/main/java/com/kwwsyk/endinv/common/client/gui/EndInvSettingScreen.java
@@ -11,7 +11,6 @@ import net.minecraft.client.gui.components.CycleButton;
 import net.minecraft.client.gui.components.EditBox;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
@@ -27,7 +26,7 @@ import static com.kwwsyk.endinv.common.client.CachedSrcInv.INSTANCE;
 import static com.kwwsyk.endinv.common.client.ClientModInfo.getClientConfig;
 public class EndInvSettingScreen extends Screen {
 
-    private static final ResourceLocation BLANK_LOCATION = new ResourceLocation("minecraft","textures/gui/demo_background.png");
+    //private static final ResourceLocation BLANK_LOCATION = new ResourceLocation("minecraft","textures/gui/demo_background.png");
     private static final int CONFIG_ENTRY_Y_OFFSET = 17;
     private static final int CONFIG_ENTRY_X_OFFSET = 10;
     private static final int ENTRY_HEIGHT = 20,MAX_ENTRY_COUNT = 7;