+++ common/src/main/java/com/kwwsyk/endinv/common/network/payloads/toServer/QuickMoveToPagePayload.java
@@ -6,7 +6,6 @@ import com.kwwsyk.endinv.common.network.payloads.ModPacketPayload;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.inventory.AbstractContainerMenu;
-import net.minecraft.world.inventory.InventoryMenu;
 import net.minecraft.world.inventory.Slot;
 
 public record QuickMoveToPagePayload(int slotId) implements ModPacketPayload {
@@ -30,9 +29,9 @@ public record QuickMoveToPagePayload(int slotId) implements ModPacketPayload {
         optional.ifPresent(manager -> {
             AbstractContainerMenu menu = manager.getMenu();
             int slotId1 = slotId;
-            if(menu instanceof InventoryMenu && player.isCreative() && slotId>=45){//if player is in CreativeInventoryScreen$ItemPickerManu
+            /*if(menu instanceof InventoryMenu && player.isCreative() && slotId>=45){//if player is in CreativeInventoryScreen$ItemPickerManu
                 slotId1 = slotId - 9;
-            }
+            }*/
             Slot slot = menu.getSlot(slotId1);
             manager.slotQuickMoved(slot);
         });